backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-openai
chromadb
PyPDF2
python-docx
python-dotenv
sentence-transformers
numpy
pydantic
aiofiles
